+++ vslib/vpp/vppxlate/SaiVppXlate.h
@@ -394,7 +394,7 @@ typedef enum {
     /* SPAN (port mirroring) */
     extern int vpp_span_enable_disable(uint32_t sw_if_index_from,
                                     uint32_t sw_if_index_to,
-                                    uint8_t state, /* 0 = disable */
+                                    uint32_t state, /* 0 = disable */
                                     bool is_l2);
 
     extern int vpp_sflow_enable_disable(const char *hwif_name, bool enable);
